+++ src/bpf/bpf-skel-wrapper.h.in
@@ -7,7 +7,7 @@
  * fine given that LGPL-2.1-or-later downgrades to GPL if needed.
  */
 
-#include "bpf-dlopen.h"          /* IWYU pragma: keep */
+#include "bpf-util.h"          /* IWYU pragma: keep */
 
 /* libbpf is used via dlopen(), so rename symbols */
 #define bpf_object__attach_skeleton sym_bpf_object__attach_skeleton
