+++ src/core/bpf-socket-bind.c
@@ -9,7 +9,7 @@
 
 #if BPF_FRAMEWORK
 /* libbpf, clang, llvm and bpftool compile time dependencies are satisfied */
-#include "bpf-dlopen.h"
+#include "bpf-util.h"
 #include "bpf-link.h"
 #include "socket-bind-api.bpf.h"
 #include "socket-bind-skel.h"
