+++ src/shared/meson.build
@@ -19,8 +19,8 @@ shared_sources = files(
         'boot-entry.c',
         'boot-timestamps.c',
         'bootspec.c',
-        'bpf-dlopen.c',
         'bpf-program.c',
+        'bpf-util.c',
         'bridge-util.c',
         'btrfs-util.c',
         'bus-get-properties.c',
