+++ src/test/test-bpf-token.c
@@ -1,26 +1,34 @@
 /* SPDX-License-Identifier: LGPL-2.1-or-later */
 
-#include <bpf/bpf.h>
 #include <fcntl.h>
 
+#include "bpf-util.h"
 #include "fd-util.h"
 #include "main-func.h"
-#include "tests.h"  /* NOLINT(misc-include-cleaner): this is needed conditionally */
+#include "tests.h"
 
 static int run(int argc, char *argv[]) {
-#if defined(LIBBPF_MAJOR_VERSION) && (LIBBPF_MAJOR_VERSION > 1 || (LIBBPF_MAJOR_VERSION == 1 && LIBBPF_MINOR_VERSION >= 5))
-        _cleanup_close_ int bpffs_fd = open("/sys/fs/bpf", O_RDONLY);
+#if HAVE_LIBBPF
+        int r;
+
+        r = dlopen_bpf(LOG_ERR);
+        if (r < 0)
+                return r;
+
+        _cleanup_close_ int bpffs_fd = open("/sys/fs/bpf", O_CLOEXEC|O_RDONLY);
         if (bpffs_fd < 0)
                 return log_error_errno(errno, "Failed to open '/sys/fs/bpf': %m");
 
-        _cleanup_close_ int token_fd = bpf_token_create(bpffs_fd, /* opts= */ NULL);
+        _cleanup_close_ int token_fd = sym_bpf_token_create(bpffs_fd, /* opts= */ NULL);
+        if (token_fd == -ENOSYS)
+                return log_tests_skipped("bpf_token_create() unavailable (libbpf too old or kernel lacks BPF_TOKEN_CREATE support)");
         if (token_fd < 0)
-                return log_error_errno(errno, "Failed to create bpf token: %m");
+                return log_error_errno(token_fd, "Failed to create bpf token: %m");
 
         log_info("Successfully created token fd.");
         return 0;
 #else
-        return log_tests_skipped("libbpf is older than v1.5");
+        return log_tests_skipped("BPF framework support is disabled");
 #endif
 }
